Create empty
A Portugese bank is rolling out term deposit for its customers. They have in the past connected to their
customer base through phone calls. The goal here is to identify/predict which customers are more likely to respond to product marketing calls and only target them instead of the whole base.